Add files via upload
First version of the tool. Adaptation of the Fortran program (subroutine_add_hydrogen.f) in Python (subroutine_add_hydrogen.py).
The three functions in Python has been tested on simple examples and returned correct answers.
I tried to translate the case 3 (CH2) of the Fortran switch in Python using the previous functions.
I've done several tests with the C5 carbon of the first lipid in the start_protonateded.gro structure (generated with Gromacs tool and ffgmx2.hdb parameter file). So far, wasn't able to get correct 3D coordinates for the hydrogens. But, when I checked the structure of the lipid, there is not an angle of 109.5° between the hydrogens and N4/C5/C6 atoms. 
I assume that in the Fortran code, v1 is the central atom meaning C5 here. v2 and v3 are named helpers meaning N4 (-> v2) and C6 (->v3) in my example. 
Also, I have no effective way to test if the vectors that I used are correctly computed (using VMD for example) and doing so I'm not sure that the procedure that I've coded is correct too.